requirements.txt
aiofiles==23.2.1
aiomysql==0.2.0
altgraph==0.17.4
annotated-types==0.7.0
anthropic==0.31.2
anyio==4.4.0
async-timeout==4.0.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cssselect==1.2.0
cycler==0.12.1
distro==1.9.0
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.110.2
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.0
huggingface-hub==0.24.7
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
jieba==0.42.1
jiter==0.5.0
jmespath==1.0.1
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
modin==0.29.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pefile==2024.8.26
pillow==10.4.0
pip==24.2
playwright==1.42.0
psutil==6.0.0
pycryptodome==3.20.0
pydantic==2.5.2
pydantic_core==2.14.5
pyee==11.0.1
PyExecJS==1.5.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMySQL==1.1.1
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyrabbit==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==4.6.0
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tenacity==8.2.2
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.29.0
w3lib==2.2.1
wheel==0.44.0
wordcloud==1.9.3
zipp==3.20.1
